requirements.txt
asgiref==3.2.3
dj-database-url==0.5.0
Django==3.0.3
django-heroku==0.3.1
django-tinymce4-lite==1.8.0
gunicorn==20.0.4
jsmin==2.2.2
mysqlclient==1.4.6
Pillow==7.2.0
psycopg2==2.8.5
pytz==2019.3
shortid==0.1.2
sqlparse==0.3.0
whitenoise==5.1.0
